Ignore errors on closing sockets
Previously, when you manually closed a socket and an error fired as you
did so, it would fire the close event twice on the socket, once for the
error, and once for the actual close. This was causing the reconnection
logic to fire twice, the first time realising it was a clean close, and
the second thinking it was an error, therefor reconnecting.